Switch back to Batik compatibility in SVG font's strikethrough-position.
In commit [cf8773d9] I switched the sign convention of strikethrough-position to
make it consistent with the SVG 1.1 specification. However, SVG fonts are no longer
supported by mainstream tools, and apparently all the current SVG fonts out there
come from two sources which use the other convention:

- The Apache Batik project, somewhat inactive but still the most commonly used SVG
font software.
- A SVG font converter by company SteadyState, which is no longer available. Some
SVG font test files in Batik (and also this project) were generated with that tool.

So we are back to that convention.