CI: exclude tests from CI.
When executed on Github's CI, the SamplesRenderingTest.testFontArabic() test fails.
That test is likely to keep failing in the future, due to subtle variations in the
Windows support for its default Arabic font.

The test runs fine on my latest Windows 10.0.19043, but I already had to update the
reference image after a OS upgrade, and apparently Github uses Windows Server which
is not rendering the arabic font in the exact same way.

It is unfortunate, but tests have to be excluded for now.