Tests: update test references with more accurate images.
After the latest adjustments to replace proprietary fonts, the references being
updated by this commit did no longer accurately match what was being produced.
That was not discovered until now, as a consequence of #34.

With further adjustments and some default font modifications, the new references
are now appropriate for the new image comparison infrastructure.