bridge: be tolerant to missing references #80
Instead of throwing an exception (and stop processing) on a missing reference, just
report the error via the user agent and keep processing. This is what a conformant
user agent is supposed to do.

There are other circumstances (like other I/O errors) where it would make sense to
behave in the same way, but this is deferred to future updates.

The `slideshow` and `extension` modules are also affected.