transcoder: add the `KEY_SVG_SELECTOR` transcoding hint
If `KEY_SVG_SELECTOR` is set and the document is HTML, its value as a CSS selector
is used to locate the element that will be used as the SVG root.

In the case of the transcoding helper, if the selector argument is null then
`KEY_SVG_SELECTOR` is used, and if also null then the previous behaviour follows.

This new hint is part of the effort to make EchoSVG friendly to embedded SVG images.
See also: #40