bridge: replace the `WeakReference` to the engine in `SVGAnimationEngine` with a `SoftReference`
In `SVGAnimationEngine` the animation engine is kept as a `WeakReference`, but
with the current JVMs those references do not last very long. This may cause
some animations to not happen.

Also due to that, the `SwingMemoryLeakTest` requires at least 4GB of heap space
to pass and is triggering failures. Switching the `WeakReference` by a `SoftReference`
may seem a bit radical but is the most conservative way to fix this.

As a consequence of this change, the stress-tests for `WeakReference`d objects
(that were failing, as explained above) are no longer executed during the build.
It could be said that such tests achieved what they were intended for: check for
the validity of the `WeakReference` approach.

Closes #103.