css: handle line-height in the base SVG implementation
The SVG 1.1 implementation of Batik does not handle the line-height property,
and uses a fixed value of 1 instead (in SVG 1.1 the line height is defined to
be 1). The SVG 1.2 implementation does manage that property but uses a default
of 1.1.

This commit unifies the handling of line-height so the CSS property can be used
by both the base (SVG1.1) and 1.2 implementations, and both use the same default
(1.1 was chosen due to accessibility concerns).

This unification is important for the support of the CSS units `lh` and `rlh`
and for the handling of foreign elements.

After this commit, images rendered with the SVG 1.1 implementation of Batik may
show a different line-height with respect to EchoSVG.