Split the transcoder module in 4 modules: api, svg, svg2svg and tosvg
Sometimes an SVG library is embedded into an executable, and a common concern
for both EchoSVG and Batik users in those cases is to make the size of the
embedded modules as small as possible.

Notably, in the case of the Transcoder a number of unneeded classes may be
included. For an example of such concern see

https://github.com/eclipse-platform/eclipse.platform.swt/issues/1438#issuecomment-2454551940

The current `transcoder` module was designed as a monolithic do-it-all package,
with users always carrying code that they do not intend to run. But it can be
split in several submodules, in a way that backwards compatibility is kept (the
old `transcoder` module still exists and provides all the classes), so users
willing to minimize their dependencies can choose a smaller specific sub-module.

This commit splits the transcoder in 4 modules:

- `transcoder-api` (API)
- `transcoder-svg` (SVG to image)
- `transcoder-svg2svg` (SVG to SVG)
- `transcoder-tosvg` (conversions to SVG, currently only WMF to SVG)

In the typical case of rendering SVG as a PNG, the current 2.0 transcoder with
dependencies fills a jar of 6.26 MB, while the specific SVG-to-image transcoder
becomes a bit smaller (5.98 MB).

The gains are higher for the WMF to SVG case (3.94 MB) or the SVG-to-SVG (3.60
MB). Note that there is plenty of room for improvement in conversions to SVG,
because the current code uses the SVG DOM implementation provided by the `anim`
module (with all its dependencies) but the JDK DOM implementation could be used
instead (like in the EchoSVG Wiki example for `SVGGraphics2D`).