transcoder: deprecate method `SVGAbstractTranscoder.createBridgeContext(String)`
The `version` attribute was dropped from SVG. If you wanted to override that
method, please do that with `createBridgeContext()` and `createSVG12BridgeContext()`
instead. The former creates a `BridgeContext` which is intended for the current
SVG specification, not just 1.1 as in Batik, while the latter produces one for
the deprecated 1.2 version.