bridge: support the modern `cursor` CSS property
The current handling of the CSS property does not match the UI specification nor
the behaviour of web browsers. It relies on the `cursor` SVG element which was
removed from the specification.

Now EchoSVG keeps compatibility with the deprecated behaviour, but at the same
time supports the browser-compatible one.

See https://www.w3.org/TR/css-ui-3/#cursor