awt: use the three-arguments `ColorConvertOp` constructor
The two-arguments constructor of `ColorConvertOp` leads to a buggy code path; it
seems that this was known to the author of the relevant classes, as he used a
workaround in another class of that package (avoiding that class completely).

Stanimir Stamenkov @stanio found that the three-arguments constructor works as
intended, see BATIK-1383.